requirements.txt
astroid==2.2.5
CacheControl==0.12.5
cachetools==3.1.0
firebase-admin==2.16.0
gcloud==0.17.0
google-api-core==1.11.0
google-auth==1.6.3
google-cloud-core==1.0.0
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
grpcio==1.20.1
httplib2==0.12.3
isort==4.3.20
jws==0.1.3
lazy-object-proxy==1.4.1
mccabe==0.6.1
msgpack==0.6.1
oauth2client==3.0.0
protobuf==3.8.0rc1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycryptodome==3.4.3
pylint==2.3.1
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2019.1
requests==2.11.1
requests-toolbelt==0.7.0
rope==0.14.0
rsa==4.0
six==1.12.0
typed-ast==1.3.5
wrapt==1.11.1
